Move and register sweeper
